+++ src/main/java/org/astrojournal/generator/absgen/LatexExporter.java
@@ -40,6 +40,8 @@ import org.astrojournal.configuration.ajconfiguration.AJPropertyConstants;
 import org.astrojournal.generator.Report;
 import org.astrojournal.generator.headfoot.LatexFooter;
 import org.astrojournal.generator.headfoot.LatexHeader;
+import org.astrojournal.generator.statistics.BasicStatistics;
+import org.astrojournal.generator.statistics.LatexStatistics;
 import org.astrojournal.utilities.RunExternalCommand;
 
 /**
@@ -59,6 +61,9 @@ public abstract class LatexExporter extends Exporter {
 
     /** If the LaTeX output should be printed. */
     protected boolean latexOutput = false;
+
+    /** The LaTeX filename for storing the statistics. */
+    protected String basicStatisticsFilename = "BasicStatistics.tex";
 
     /**
      * Default constructor.
@@ -97,7 +102,7 @@ public abstract class LatexExporter extends Exporter {
      * Generate the LaTeX document sorting the observation by decreasing date.
      */
     @Override
-    public boolean generateJournal() {
+    public boolean generateJournal(BasicStatistics basicStatistics) {
 	LatexHeader latexHeader = new LatexHeader(filesLocation,
 		headerFooterFolder, headerFilename);
 	LatexFooter latexFooter = new LatexFooter(filesLocation,
@@ -108,7 +113,7 @@ public abstract class LatexExporter extends Exporter {
 	    writer = new BufferedWriter(new OutputStreamWriter(
 		    new FileOutputStream(filesLocation + File.separator
 			    + reportFilename), "utf-8"));
-	    writeLatexMain(writer, latexHeader, latexFooter);
+	    writeLatexMain(writer, latexHeader, latexFooter, basicStatistics);
 
 	} catch (IOException ex) {
 	    log.error("Error when opening the file " + filesLocation
@@ -141,10 +146,13 @@ public abstract class LatexExporter extends Exporter {
      * @param writer
      * @param latexHeader
      * @param latexFooter
+     * @param basicStatistics
+     *            The statistics to write
      * @throws Exception
      */
     public abstract void writeLatexMain(Writer writer, LatexHeader latexHeader,
-	    LatexFooter latexFooter) throws Exception;
+	    LatexFooter latexFooter, BasicStatistics basicStatistics)
+	    throws Exception;
 
     /**
      * This method contains the LaTeX code for the document content.
@@ -155,6 +163,62 @@ public abstract class LatexExporter extends Exporter {
      */
     public abstract void writeLatexContent(Writer writer, Report report)
 	    throws IOException;
+
+    /**
+     * Write the statistics section.
+     * 
+     * @param writer
+     *            the Writer
+     * @throws Exception
+     */
+    public void writeSectionStatistics(Writer writer) throws Exception {
+	writer.write("\\clearpage\n");
+	writer.write("\\section{Basic Statistics}\n");
+	// include the file removing the extension .tex
+	writer.write("\\begin{center} \n");
+	writer.write("\\input{" + reportFolder + "/"
+		+ basicStatisticsFilename.replaceFirst("[.][^.]+$", "") + "}\n");
+	writer.write("\\end{center} \n");
+    }
+
+    /**
+     * Write the statistics to a LaTeX file.
+     * 
+     * @param basicStatistics
+     *            The statistics to write
+     * 
+     * @return true if the file was written correctly
+     */
+    public boolean writeLatexStatistics(BasicStatistics basicStatistics) {
+	Writer writer = null;
+	try {
+	    writer = new BufferedWriter(new OutputStreamWriter(
+		    new FileOutputStream(new File(filesLocation
+			    + File.separator + reportFolder,
+			    basicStatisticsFilename)), "utf-8"));
+	    LatexStatistics latexStatistics = new LatexStatistics();
+	    latexStatistics.writeAll(writer, basicStatistics);
+	} catch (IOException ex) {
+	    log.error("Error when opening the file " + filesLocation
+		    + File.separator + reportFolder + File.separator
+		    + basicStatisticsFilename);
+	    log.debug("Error when opening the file " + filesLocation
+		    + File.separator + reportFolder + File.separator
+		    + basicStatisticsFilename, ex);
+	    return false;
+	} catch (Exception ex) {
+	    log.debug(ex);
+	    log.error(ex, ex);
+	    return false;
+	} finally {
+	    try {
+		if (writer != null)
+		    writer.close();
+	    } catch (IOException e) {
+	    }
+	}
+	return true;
+    }
 
     @Override
     public void postProcessing() throws IOException {